Move exception handling to a new class
PresentScreenings/Main.cs:
Move exception handling to a new class.

PresentScreenings/PresentScreenings.csproj:
Add class AlertRaiser.

PresentScreenings/Utilities/AlertRaiser.cs:
Add a new class to handle fatal exceptions.
