+++ PresentScreenings/PresentScreenings.csproj
@@ -171,6 +171,7 @@
     <Compile Include="Utilities\ScreeningsPlanner.cs" />
     <Compile Include="Utilities\WebUtility.cs" />
     <Compile Include="Table Classes\FilmOutlineLevel.cs" />
+    <Compile Include="Utilities\AlertRaiser.cs" />
   </ItemGroup>
   <ItemGroup>
     <InterfaceDefinition Include="Main.storyboard" />
